+++ installers/install.ps1
@@ -30,7 +30,7 @@ $InstallMethod = if ($ScriptDir -and ((Test-Path "$ScriptDir\ccs.ps1") -or (Test
 # IMPORTANT: Update this version when releasing new versions!
 # This hardcoded version is used for standalone installations (irm | iex)
 # For git installations, VERSION file is read if available
-$CcsVersion = "2.2.2"
+$CcsVersion = "2.2.3"
 
 # Try to read VERSION file for git installations
 if ($ScriptDir) {